requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
Markdown==3.7
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2024.2
